requirements.txt
streamlit==1.29.0
pydantic==2.5.0
python-dotenv==1.0.0
ollama==0.1.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
pdfplumber==0.10.3
duckduckgo-search==4.1.1
requests==2.31.0
beautifulsoup4==4.12.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
pandas==2.1.3
pillow==10.1.0